raw data
Inpatient Data: Provides the details of the claims filed for the patients who were admitted to the hospitals for their treatment.
Outpatient Data: Provides the details related to the claims filed for the patients who visited the hospitals but were not admitted.
Beneficiary Data: Contains the demographic, disease, premium, and reimbursement related information of the beneficiaries/patients.
label Data: Contains the mapping of Provider’s unique id and the class label signifying whether the Provider is fraud or not. For the Test data, only the Provider’s unique id is given and the class label has to be found.